# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Lifestyle Super App API
  description: |
    ## 🚀 Lifestyle Super App - REST API Documentation
    
    **Base URL:** `https://api.lifestyle.vn/v1`  
    **Local Dev:** `http://localhost:3000/api/v1`
    
    ### Authentication
    Hầu hết endpoints yêu cầu JWT Bearer token:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ### Response Format
    ```json
    {
      "success": true,
      "data": { ... },
      "message": "Success",
      "meta": { "page": 1, "limit": 20, "total": 100 }
    }
    ```
    
    ### Error Format
    ```json
    {
      "success": false,
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Email is required",
        "statusCode": 400
      }
    }
    ```
    
    ### Modules
    | Module | Prefix | Description |
    |--------|--------|-------------|
    | Auth | `/auth` | Đăng ký, Đăng nhập, Token |
    | Users | `/users` | Hồ sơ, Địa chỉ, Preferences |
    | Drivers | `/drivers` | Đăng ký tài xế, Xe, Giấy tờ |
    | Merchants | `/merchants` | Cửa hàng, Nhân viên |
    | Wallets | `/wallets` | Ví, Nạp tiền, Rút tiền |
    | Payments | `/payments` | Thanh toán, Webhooks |
    | Spotlight | `/spotlight` | Redcomments, Reviews, Creators |
    | Loyalty | `/loyalty` | Xu, Referral |

  version: 1.0.0
  contact:
    name: Lifestyle Super App Team
    email: dev@lifestyle.vn
  license:
    name: Private
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development
...
security:
  - BearerAuth: []

# =============================================================================
# PATHS (API Endpoints)
# =============================================================================
paths:
...
  /spotlight/reels:
    get:
      tags: [Spotlight]
      summary: Lấy danh sách video Redcomment (TikTok-style feed)
      description: |
        Feed dạng infinite scroll, trả về Redcomments dạng video reel.
        Sắp xếp theo thuật toán (views + recency + relevance).
      security: []
      parameters:
        - name: category
          in: query
          description: Lọc theo danh mục
          schema:
            type: string
            enum: [PRODUCT, RESTAURANT, CAFE, HOTEL, TRAVEL, INSURANCE, EVENT, EXPERIENCE]
        - name: format
          in: query
          schema:
            type: string
            enum: [VIDEO_REEL, ARTICLE, PHOTO_ESSAY, COMPARISON]
            default: VIDEO_REEL
        - name: cursor
          in: query
          description: Cursor cho infinite scroll (ID của item cuối cùng)
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Danh sách Redcomments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReelsFeedResponse'
              example:
                success: true
                data:
                  - id: "uuid-rdc-1"
                    redcommentNumber: "LS-RDC-260216-001"
                    creator:
                      id: "uuid-creator-1"
                      displayName: "Trần Thị B - Food Reviewer"
                      avatarUrl: "https://cdn.lifestyle.vn/avatars/creator1.jpg"
                      tier: "RISING"
                      followerCount: 5200
                    format: "VIDEO_REEL"
                    title: "Review Phở Nguyễn - Phở Bò NGON NHẤT Quận 1!"
                    description: "Nước dùng trong, ngọt thanh, thịt bò tươi mềm."
                    videoUrl: "https://cdn.lifestyle.vn/spotlight/video1.mp4"
                    videoDuration: 45
                    thumbnailUrl: "https://cdn.lifestyle.vn/spotlight/thumb1.jpg"
                    targetType: "RESTAURANT"
                    targetName: "Phở Nguyễn"
                    overallRating: 4.5
                    views: 12500
                    likes: 850
                    shares: 120
                    commentsCount: 45
                    isLiked: false
                    isSaved: false
                    ctaButtons:
                      - text: "Đặt ngay"
                        action: "BUY"
                        targetUrl: "lifestyle://merchant/uuid-merchant-1"
                    publishedAt: "2026-02-13T14:30:00Z"
                nextCursor: "uuid-rdc-1"
                hasMore: true
  /spotlight/redcomments:
    post:
      tags: [Spotlight]
      summary: Tạo Redcomment mới (Creator only)
      description: |
        KOC/Creator tạo nội dung mới. Hỗ trợ 4 format:
        - VIDEO_REEL (upload video vertical)
        - ARTICLE (bài viết)
        - PHOTO_ESSAY (bộ ảnh)
        - COMPARISON (so sánh)
        
        Status mặc định: DRAFT. Creator submit để chuyển sang PENDING_REVIEW.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateRedcommentRequest'
      responses:
        '201':
          description: Redcomment đã tạo (DRAFT)
          content:
            application/json:
              example:
                success: true
                data:
                  id: "uuid-rdc-new"
                  redcommentNumber: "LS-RDC-260216-002"
                  status: "DRAFT"
                  message: "Redcomment đã lưu nháp. Submit để gửi duyệt."
  /spotlight/redcomments/{redcommentId}:
    parameters:
      - name: redcommentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    
    get:
      tags: [Spotlight]
      summary: Lấy chi tiết Redcomment
      security: []
      responses:
        '200':
          description: Chi tiết Redcomment
    
    patch:
      tags: [Spotlight]
      summary: Cập nhật Redcomment (Creator only, khi còn DRAFT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                targetTags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Đã cập nhật
...
  /spotlight/reviews:
    post:
      tags: [Spotlight]
      summary: Viết đánh giá (User)
      description: User viết review cho sản phẩm/dịch vụ đã sử dụng.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review đã gửi, chờ duyệt
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token. Lấy từ `/auth/login` hoặc `/auth/register`.
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIs...
        ```
    SepayWebhookKey:
      type: apiKey
      in: header
      name: X-Sepay-Key
      description: API Key xác thực webhook từ SePay

  # ---------------------------------------------------------------------------
  # Parameters (Reusable)
  # ---------------------------------------------------------------------------
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Số item mỗi trang
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    CreateRedcommentRequest:
      type: object
      required: [format, title, targetType, targetName, overallRating]
      properties:
        format:
          type: string
          enum: [VIDEO_REEL, ARTICLE, PHOTO_ESSAY, COMPARISON]
        title:
          type: string
          maxLength: 300
        description:
          type: string
          maxLength: 500
        video:
          type: string
          format: binary
          description: Video file (cho VIDEO_REEL)
        photos:
          type: array
          items:
            type: string
            format: binary
          description: Ảnh (cho PHOTO_ESSAY)
        coverImage:
          type: string
          format: binary
        targetType:
          type: string
          enum: [PRODUCT, RESTAURANT, CAFE, HOTEL, TRAVEL, INSURANCE, EVENT, EXPERIENCE]
        targetId:
          type: string
          format: uuid
        targetName:
          type: string
        targetTags:
          type: array
          items:
            type: string
        merchantId:
          type: string
          format: uuid
        overallRating:
          type: number
          minimum: 1
          maximum: 5
        ctaButtons:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              action:
                type: string
                enum: [BOOK, CONSULT, BUY, CALL, EXTERNAL_LINK]
              targetUrl:
                type: string
    ReelsFeedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              redcommentNumber:
                type: string
              creator:
                type: object
              format:
                type: string
              title:
                type: string
              videoUrl:
                type: string
              thumbnailUrl:
                type: string
              targetType:
                type: string
              overallRating:
                type: number
              views:
                type: integer
              likes:
                type: integer
              isLiked:
                type: boolean
              ctaButtons:
                type: array
                items:
                  type: object
        nextCursor:
          type: string
        hasMore:
          type: boolean
    CreateReviewRequest:
      type: object
      required: [targetType, targetName, overallRating, body]
      properties:
        targetType:
          type: string
          enum: [PRODUCT, RESTAURANT, CAFE, HOTEL, TRAVEL, INSURANCE, EVENT, EXPERIENCE]
        targetId:
          type: string
          format: uuid
        targetName:
          type: string
        merchantId:
          type: string
          format: uuid
        overallRating:
          type: number
          minimum: 1
          maximum: 5
        title:
          type: string
          maxLength: 200
        body:
          type: string
        photos:
          type: array
          items:
            type: string
            format: binary
        video:
          type: string
          format: binary